refactor(AutoOperate): extracting common interfaces for AutoOperate related components (GitHub #166)
