+++ crane4j-core/src/main/java/cn/crane4j/core/support/aop/AutoOperateAnnotatedElementResolver.java
@@ -1,31 +1,9 @@
 package cn.crane4j.core.support.aop;
 
 import cn.crane4j.annotation.AutoOperate;
-import cn.crane4j.core.exception.Crane4jException;
-import cn.crane4j.core.executor.BeanOperationExecutor;
-import cn.crane4j.core.parser.BeanOperationParser;
-import cn.crane4j.core.parser.BeanOperations;
-import cn.crane4j.core.parser.operation.KeyTriggerOperation;
-import cn.crane4j.core.support.Crane4jGlobalConfiguration;
-import cn.crane4j.core.support.Grouped;
-import cn.crane4j.core.support.MethodInvoker;
-import cn.crane4j.core.support.TypeResolver;
-import cn.crane4j.core.support.reflect.PropertyOperator;
-import cn.crane4j.core.util.Asserts;
-import cn.crane4j.core.util.ClassUtils;
-import cn.crane4j.core.util.CollectionUtils;
-import cn.crane4j.core.util.StringUtils;
-import lombok.RequiredArgsConstructor;
-import org.checkerframework.checker.nullness.qual.NonNull;
+import org.checkerframework.checker.nullness.qual.Nullable;
 
 import java.lang.reflect.AnnotatedElement;
-import java.lang.reflect.Method;
-import java.lang.reflect.Parameter;
-import java.util.HashSet;
-import java.util.Optional;
-import java.util.Set;
-import java.util.function.Function;
-import java.util.function.Predicate;
 
 /**
  * <p>An handler that resolve the {@link AutoOperate}
@@ -33,12 +11,10 @@ import java.util.function.Predicate;
  *
  * @author huangchengxing
  * @see AutoOperate
+ * @see AutoOperateAnnotatedElement
+ * @see MethodBasedAutoOperateAnnotatedElementResolver
  */
-@RequiredArgsConstructor
-public class AutoOperateAnnotatedElementResolver {
-
-    private final Crane4jGlobalConfiguration configuration;
-    private final TypeResolver typeResolver;
+public interface AutoOperateAnnotatedElementResolver {
 
     /**
      * Resolve the {@link AutoOperate} annotation on the element
@@ -48,88 +24,6 @@ public class AutoOperateAnnotatedElementResolver {
      * @param annotation annotation
      * @return {@link AutoOperateAnnotatedElement}
      */
-    public AutoOperateAnnotatedElement resolve(@NonNull AnnotatedElement element, @NonNull AutoOperate annotation) {
-        Asserts.isNotNull(element, "element must not be null");
-        Asserts.isNotNull(annotation, "annotation must not be null");
-
-        MethodInvoker extractor = resolveExtractor(element, annotation);
-        // prepare components for use
-        BeanOperationParser parser = configuration.getBeanOperationsParser(annotation.parser(), annotation.parserType());
-        BeanOperationExecutor executor = configuration.getBeanOperationExecutor(annotation.executor(), annotation.executorType());
-        Predicate<? super KeyTriggerOperation> filter = resolveFilter(annotation);
-
-        // type is specified in annotation
-        Class<?> annotationSpecifiedType = annotation.type();
-        if (!ClassUtils.isObjectOrVoid(annotationSpecifiedType)) {
-            BeanOperations beanOperations = parser.parse(annotation.type());
-            return AutoOperateAnnotatedElement.forStaticTypeOperation(
-                annotation, element, extractor, filter, beanOperations, executor
-            );
-        }
-
-        // type not specified in annotation, but element is class
-        // so use itself as type as operation source
-        if (element instanceof Class) {
-            BeanOperations beanOperations = parser.parse(element);
-            return AutoOperateAnnotatedElement.forStaticTypeOperation(
-                annotation, element, extractor, filter, beanOperations, executor
-            );
-        }
-
-        // not specify type, delay parsing until execution time
-        Function<Object, BeanOperations> dynamicParser = t -> Optional.ofNullable(t)
-            .map(typeResolver::resolve)
-            .map(parser::parse)
-            .orElse(BeanOperations.empty());
-        return AutoOperateAnnotatedElement.forDynamicTypeOperation(
-            annotation, element, extractor, filter, executor, dynamicParser
-        );
-    }
-
-    private MethodInvoker resolveExtractor(AnnotatedElement element, AutoOperate annotation) {
-        Class<?> type = resolveTypeForExtractor(element);
-        String on = annotation.on();
-        MethodInvoker extractor = (t, args) -> t;
-        if (StringUtils.isNotEmpty(on)) {
-            PropertyOperator propertyOperator = configuration.getPropertyOperator();
-            extractor = propertyOperator.findGetter(type, on);
-            Asserts.isNotNull(extractor, "cannot find getter for [{}] on [{}]", on, annotation.type());
-        }
-        return extractor;
-    }
-
-    private static Class<?> resolveTypeForExtractor(AnnotatedElement element) {
-        if (element instanceof Method) {
-            return ((Method)element).getReturnType();
-        } else if (element instanceof Parameter) {
-            return ((Parameter)element).getType();
-        } else if (element instanceof Class) {
-            return (Class<?>)element;
-        } else {
-            throw new Crane4jException("cannot resolve type for annotated element [{}]", element);
-        }
-    }
-
-    /**
-     * Resolve group for {@link AutoOperate#includes()} and {@link AutoOperate#excludes()}.
-     *
-     * @param annotation annotation
-     * @return actual include groups
-     */
-    protected static Predicate<? super KeyTriggerOperation> resolveFilter(AutoOperate annotation) {
-        Set<String> excludes = CollectionUtils.newCollection(HashSet::new, annotation.excludes());
-        Set<String> includes = CollectionUtils.newCollection(HashSet::new, annotation.includes());
-        includes.removeAll(excludes);
-        // nothing includes
-        if (includes.isEmpty()) {
-            return Grouped.noneMatch(annotation.excludes());
-        }
-        // nothing excludes
-        if (excludes.isEmpty()) {
-            return Grouped.anyMatch(annotation.includes());
-        }
-        // include or not exclude
-        return t -> CollectionUtils.containsAny(includes, t.getGroups())
-            || CollectionUtils.notContainsAny(excludes, t.getGroups());
-    }
+    AutoOperateAnnotatedElement resolve(
+        AnnotatedElement element, @Nullable AutoOperate annotation);
 }
